FDG-10264 Select all
