+++ src/components/data-preview/data-preview-chart/chart-legend-panel/chart-legend-panel.module.scss
@@ -14,7 +14,7 @@
 
 .buttonContainer {
   padding: 1rem;
-  max-height: 373px;
+  max-height: 385px;
   overflow-y: auto;
 }
