+++ src/components/data-table/column-select/data-table-column-selector.module.scss
@@ -45,7 +45,7 @@
   justify-content: space-between;
   align-items: center;
   border-bottom: solid 1px #d9d9d9;
-
+  padding: 0.375rem 0;
   .selectAllColumns {
     border-bottom: 0;
   }
