+++ src/components/select-all/select-all.module.scss
@@ -78,5 +78,5 @@ $SelectAllWidth: 261px;
 }
 
 .container {
-  padding: 1rem;
+  padding: 0.625rem 1rem;
 }
